Fix scaling anchors in PMv2 applications. (#8349)
* Cache DisplayRect used to compute anchors

Fixes #8266.

In PermonV2 mode applications, when moved from one monitor to the other, Form/Container is scaled with scale factor computed with respect to new DPI on the monitor. However, scaling ratio between non-cleant and client area is not linear. This is causing few pixel variance when anchors are also scaled with same scale factor.
Anchors are relative to display rectangle. Hence, computing anchor scale factor with repsect to chnage in Display rect instead of chnage in the bounds (a.k.a: scale factor computed with respect to new DPI).

* Address feedack.