Fix anchors in nested UserControls (#8201)
* Fix anchored nested controls in the UserControl hosted on the Form.

This is not a regression but the root cause is different in Framework vs <NET 8.0.
Anchor calculations are delayed untill after the Handles are created. This is problematic in the scenarios where UserControls containing anchored controls are nested.

ex:

 User control has a TextBox anchored in all directions.
 User control is then resized on Form,  TextBox anchors should have been calculated based on the UserControls original size and grow/shrink on this resize event.
 However, none of the Handles are created yet and thus anchor calculations, when happen, will be done with thi snew size for UserControl.

In this change we calculate anchors when parent's layout is resumed and not wait until handles are created or not. When handles created, and DPI applied, we scale controls as well as anchors.


Fixes #6381.

* Cleanup Handle created checks.

* Fix tests and build failures.

* Fix design-time scenario.

* Address review feedback.

* Address feedback.

* Address feedback.

* Add argument names.