Fix readonly static member evaluation order
There is a problem with the parsing of the TargetFramework, which is causing an invalid value to be set. static nullable private property needs to be assigned before its use.

Because the AppContext flags are read-only for .NET apps, and the incorrect TargetFramework is bypassing the logic behind these flags.

Found while verifying #6381.