Remove Drawing.Common .NETCoreApp PNSEs (dotnet/runtime#80435)
System.Drawing.Common was throwing PNSEs on non Windows platforms when
accessing public API which makes sense as the library isn't supposed to
be used on other platforms.

That said, we already have a runtime hook that throws when DllImports
are invoked. We intentionally want to allow accessing Drawing exchange
types to not break customers when they don't rely on GDI+ or other
native functions.

This is also part of a wider effort that allows this library to be
ported to another repository (which doesn't support targeting RIDs).

Commit migrated from https://github.com/dotnet/runtime/commit/7c265c396e6ef31cb0d588853834c800e40ecdef