Fix exception dialog handling. (#8695)
Test fixtures do not run on the same thread the test method executes on when marked with `[WinFormsFact]` or `[StaFact]`. This change removes the test fixture and adds a new context switch to throw thread exceptions.

Initialization of the switch is done programmatically via a module initializer. The module initializer will also set the starting cursor position to avoid spurious errors in tests due to cursor positioning.

The Cursor tests were the ones modifying the cursor position and causing intermittent errors in other tests. Moved them to their own collection and forced them to run sequentially.

Cleaned up WindowsFormsSynchronizationContext.

Simplified checks in LocalAppContextSwitches. We shouldn't be conditioning on supported OS beyond checking for UI availability. We don't support Windows 7, but we won't deliberately cause WinForms apps to crash on Windows by calling later APIs without checking for support. We will not add additional complexity to the code beyond that for unsupported OS platforms.