Fix disabled text label rendering.
The problem was actually in ControlPaint.DrawStringDisabled that takes an IDeviceContext. In the big refactoring it lost the application of TextFormatFlags to decide whether or not to apply Graphics state. LinkLabel never set those flags historically so it didn't get the clipping it applied to the Graphics object applied to the HDC. Fix is to apply the flags properly.

I've also cleaned up the LinkLabel code and did a thorough pass of the TextExtensions to bring the code up to snuff.