Add AppContext switch scope functionality
Adds the `BinaryFormatterScope` ref struct for explicitly setting `BinaryFormatter` enabled state.

Ran through System.Windows.Forms.Tests with the `BinaryFormatter` explicitly disabled to identify tests that were dependent and add the context to said tests.

The intent is to follow this change with additional changes to do the same for the other test assemblies. I'll be using what I find here to open additional issues or add additional context to existing ones. For example, it wasn't clear that CSV text uses the `BinaryFormatter` for the clipboard until I ran through this set of tests.

After all assemblies are scoped I'll be then looking to close test gaps.