Add BufferScope for handling buffer rentals for interop
`BufferScope` is a ref struct that handles rental and return from the `ArrayPool` as well as:

- Initial stack allocated buffers
- Reacquisition of buffers (with content copying if necessary)
- Span semantics (converters, range operators, etc.)

It significantly simplifies creating temporary buffers, particularly in cases where you need to make multiple calls (for the buffer not being large enough).

Also add `[SkipLocalsInit]` in a few places (so the stack allocated buffer doesn't get zeroed out).

Did a global removal of unused usings.