Create BinaryFormatter replacement for known types (#9088)
In a number of cases we know what data we're serializing or deserializing and can avoid using the BinaryFormatter if we weren't concerned about compatibility.

I was considering using a different format on the wire and writing a parser to deal with legacy data. It finally occured to me that if I'm going to parse legacy data I might as well put the data out in the legacy format to maintain compatibility.

This creates a model for the NRBF format that can be used to look at data the BinaryFormatter puts out. The top level of the model is represented by the BinaryFormattedObject class.

The binary records that are part of the model all self serialize and deserialize, which allows manually constructing known models to export data that can be read by the BinaryFormatter. The records never hydrate any complex types other than strings, that must be done manually.

BinaryFormatReader and BinaryFormatWriter have helper methods for creating and reading a set of known types. I've handled primitive types and Hashtable and started on List<T> as well.

Clipboard will no longer use the BinaryFormatter for primitive types. Tests have been updated appropriately.

My next step is to change the ActiveX property bag serialization to use the Hashtable serialization that this change introduces.  I also have a few more tests to port from my prototype repo.

The idea is to then further reduce the need to fall back on the BinaryFormatter by providing support for other BinaryFormatted types in WinForms and other types that WinForms projects would commonly use as public property types. This includes things like PointF and RectangleF, primitive arrays, and other collections of primitive types (List<T> and ArrayList are the priorities there).

BinaryFormattedObject does not yet support multidimensional or jagged arrays.