Move away from BinaryFormatter for ActiveX purposes (#9104)
The PropertyBags now serialize without it where possible. This allows a standard Control to completely serialize without the BinaryFormatter enabled.

For the AxHost case, I've left the BinaryFormatter as a fall-back.

I've also extended the Hashtable serialization to handle all primitive types. Note the two line change to make this happen. :)

This change also moves the "Reader" methods to "BinaryFormattedObjectExtensions". Due to our layering and the need to check multiple states I realized that we need to have all of the reading of objects directly using an existing BinaryFormattedObject instance.

Fixes a few bugs and adds more tests. Some random cleanup related to VARIANT (as I was exploring the scope of things).