Add BinaryArrayType
Sometimes single dimensional arrays get serialized as BinaryArrayType instead of ArraySingleObject. This change implements reading of BinaryArrayType for this case (writing is still not supported).

This also moves some helper methods up to ClassRecord.

It is also possible for Ids to be negative. They are effectively treated as positive numbers by the BinaryFormatter so Id was changed to do the same (but retain the original value).

Adds a batch of additional tests and moves the BinaryFormat test extensions to the test utilities assembly.