Make AccessibleObject owner references weak references (#9164)
This is to help ensure that the owner references can be be collected if the AccessibleObject CCW is the only thing still alive. We fall back onto "default" behavior if the related component is gone.

Added an internal IAccessibleOwner interface to allow attaching behavior to different AccessibleObject hierarchies. Changed most usages to TryGetOwnerAs to make it easier to audit Owner access.