Convert more User32 code to CSWin32
The converts another block of User32 interop to CSWin32.

We had created enums for messages and style flags that were specific to controls. I've decide to not recreate them as it has started to obfuscate that the values overlap with other values.

I'm still giving thought to what we do with WM_. It was convenient for debugging to have an enum, but I think we can solve that in other ways.