Convert the rest of User32 to CsWin32
This converts the rest of the User32 PInvokes. I still need to move the WM enum and DC scope, and will follow up with that.

Also removes most of NativeMethods.cs. In doing this I also tweaked the thread exception dialog to deal with the fact that CodeBase is depreciated. As assemblies don't have to be coming from disk, the file version doesn't help much either (and is much less useful than it was on Framework). Had to tweak the string to make this work.