Remove Interop.User32 class (#9331)
This change fully removes the Interop.User32 class. This does this by:

- Removing WM
- Moves GetDCScope and RegisteredMessage to Windows.Win32

Using an enum had the advantage of seeing relatively easily what a given message was (presuming it was a standard WM_ message). To help a bit with this I've created a MessageId struct that allows seeing the message description in the debugger.

One nice benefit of this change is that we now have links to the docs for the given message defines, which will help development a lot.

Also fix an intermittent test AV