Fix ActiveXImpl.InPlaceActivate
The `UIWindow` can be null, we need to check for that before trying to wrap it in an `AgileComPointer`. This was an oversight- we check to see if the object is null in the rest of the method. Also checking `pFrame` as it is possible we would fall out before trying to use it.

Also assert that the method succeeded for our own debugging purposes. Historically the result wasn't checked.

Lastly add a missing `AddRef()` and use `ThrowOnFailure()` instead of calling `Marshal` directly (for consistency).