Cleanup for utilization of target new (#9527)
Tweaks the editor config settings to encourage behavior that matches our coding style:

- We prefer `Foo foo = new()` over `var foo =`.
- `var` for built in types is an error
- Expression bodied members are always ok

A number of the rules got added by just using the editor. I'll continue to iterate.

Changed a big chunk of the code to use target typed new using fixers and regexe's for the majority of the changes.