Fix a few more ref counts in ActiveXImpl (#9634)
This is the last audit change of ActiveXImpl. Changes were vetted against docs and MFC implementations.

- QuickActivate was releasing pointers it shouldn't have been.
- SetAdvise and GetAdvise were not AddRef'ing appropriately.
- Update comments.

Getting the current thread failed in one the test legs, likely due to some memory corruption. It isn't necessary, so I've removed it to try and push the error elsewhere (where it might be more actionable). This is not a fallout of the ActiveX changes.