Release COM object when wrapping in PictureWrapper
We need to release the incoming pointer when returning a wrapped object. This was fixed in .NET 7 and removed completely in .NET 8 (where we access the COM object directly).

Without this any time we access IPicture we will permanently leak the incoming pointer with no way for customers to fix it. This usually manifests as an exhaustion of GDI handles.