Clean up IReflect in AccessibleObject a bit
Made a pass through AccessibleObject to clean up the comments a bit. Tried to clarify what was going on with IReflect.Invoke and add a few tests to characterize the current behaviors.

Also add the il file to the accessibility project so it can be browsed from the solution explorer..