Simplify Cursor code (#10005)
Construction of Cursors statics had gotten a bit over complicated. ToString() was also very expensive when creating Cursors from handles.

The core of the change is to pass the property name when constructing a Cursor from Cursors for consistency and performance.

These changes will also allow making the CursorConverter reflection free in the future.