Use ITypeInfo directly in StandardDispatch (#10254)
CreateStdDispatch doesn't do anything other than pass through to ITypeInfo so we'll do the same directly and save some of the overhead and make the code clearer.

To avoid loading and unloading the ITypeLib repeatedly we'll simply keep it alive (via the ITypeInfo) once we've loaded it once. We can expect that in an app that uses these we'll continue to need them throughout the lifetime of the app.

This allows removing the IWrapperInitialize interface, which further simplifies the way things work. I'll follow up with a change that simplifies our ComWrappers a bit based on this.