Add a project for testing trimming (#10274)
Adds a simple sample project for testing trimming. It has a TextBox, SplitContainer, Button, and ComboBox. It accepts text drops into the TextBox and uses the ClipBoard class to copy text out of the ComboBox.

This is meant to be a starting point to assess and fix the most important warnings / issues with trimming.