Move && and || to front of lines
This changes some of the code (primarily DataGrid) to match the desired code style with line breaking on long logical statements.

To align conceptualy the logical operators should be at the beginning of the line when breaking long statements.

When considering how to break lines, logical conditions grouped by a specific set of parenthesis should either all be broken out into separate lines or kept together on a single line. To show nesting, statements should be indented to match the parenthetical nesting, but parenthesis should not get their own line.

Some examples:

``` C#
    if (_trackColumn != previousColumnIndex
            && !(previousColumnIndex == -1 && hti._col == _trackColumn)
            && (dataGridViewColumnNext is null || _trackColumn != dataGridViewColumnNext.Index))
```

And:

``` C#
if (Focused
    && !IsCurrentCellInEditMode
    && (EditMode == DataGridViewEditMode.EditOnEnter
        || (EditMode != DataGridViewEditMode.EditProgrammatically && CurrentCellInternal.EditType is null)))
```

In general, if pattern matching is suggested by the IDE it should be used. `or` and `and` should go to the front of the line as well.