Move common dialogs to CsWin32 and clean printing
We can't (yet) directly use CsWin32 directly for the common dialogs due to packing differences between 32 and 64 bit. We target AnyCPU and CsWin32 doesn't have the ability to force a particular platform definition. For now, I'm copying in CsWin32 defines from a platform targetted project.

Hopefully we'll eventually get .NET support for specifying packing per platform. Also investigating getting a feature added to CsWin32 to mitigate this for us.

This does a pass through all of the print code to clean to style guides.

Plan is to follow up with an attempt to update System.Drawing to CsWin32 code.