Prototype shared CsWin32 with System.Drawing.Common
This is a rough prototype of a shared interop with System.Drawing.Common. It introduces a new base assembly for shared CsWin32 and other low-level support code.

Converted HGLOBAL, DEVMODE and related stuff to the new shared pattern. Also used the CsWin32 generated GdiplusStartup.

I have almost no idea if packaging is correct here. Assuming it isn't.

`crefs` don't work right when there is overlap in internals. #10524 tracks this, and this change disables the relevant warnings.

I've only smoke tested this with the WinFormsControlsTest.