Prototype shared CsWin32 with System.Drawing.Common (#10525)
This change adds a shared interop assembly with `System.Drawing.Common`. It introduces a new base assembly for shared CsWin32 and other low-level support code called `System.Private.Windows.Core`. This assembly is part of the `System.Drawing.Common` package.

Converted `HGLOBAL`, `DEVMODE` and related stuff to the new shared pattern. Also used the CsWin32 generated `GdiplusStartup`.

`crefs` don't work right when there is overlap in internals. #10524 tracks this, and this change disables the relevant warnings. 

Smoke tested this with the WinFormsControlsTest.