Remove Control ActiveX TraceSwitch
A significant amount of the code in some areas was just trace statements. We've removed them in other places as we have had trace breakpoints for nearly 20 years. The code is significantly easier to navigate without these.

In this particular case there also was an AV in debug for our NativeHost sample.

Also some minor cleanup, including a change to just return E_NOTIMPL from ActiveXImpl.DoVerb instead of throwing it.