Clarify ImageConverter (#10716)
ImageConverter has some conversion code with somewhat magic numbers in it. This change documents the format it's looking for and validates it more throroughly.

It also adds a new high-performance Span reader class for scenarios like this that allow bounds checked iterating through data. It makes code written for this much easier to visually validate as opposed to slicing through spans manually.

Also:

- Use SpanReader in Icon parsing code
- Fix ordering of class keywords in a number of files
- Add readonly modifier to some fields
- Minor comment cleanup
- Move ICONDIR and ICONDIRENTRY to interop folder
- Delete ENHMETAHEADER
- Move SpanHelpers
- Move POINTS to Core
- Delete unused  code in ARGB
- Move method from ARGB to ArgbBuffer