Type primitive arrays in BinaryFormattedObject (#10738)
ArraySinglePrimitive was taking IReadOnlyList<object> which necessitated boxing all items. This converts ArrayRecord to be typed so we can avoid this and adds special optimizations for byte[]. This is important for performance of the types we're supporting or intending to support (such as ImageListStreamer).

It may be possible to further improve this. If the array data is always aligned in the input then we can just read/write the data directly as byte arrays. DateTime, TimeSpan, and decimal need individually parsed.