Initial implementation of GDI+ effects
This exposes most of the GDI+ effects under the [RequiresPreviewFeatures] while we take it to .NET API review.

Effects are added under the `Imaging\Effects` namespace. The new effects are:

- BlackSaturation
- Blur
- BrightnessContrast
- ColorBalance
- ColorLookupTable
- ColorMatrix
- Contrast
- Density
- Exposure
- Highlight
- Levels
- Midtone
- Shadow
- Sharpen
- Tint
- WhiteSaturation

This also adds:

- New, efficient constructor to ColorMatrix that takes a span of floats
- Overload for Graphics.DrawImage that allows you to draw with an effect (and fix the perf of the indexer)
- An ApplyEffect method on Bitmap that lets you apply an effect to an existing bitmap
- Adds the uses preview features flag to the scratch projects

Effects are applied to bitmaps, not associated with them. The image doesn't retain a handle, it just modifies the bits using the effect. As such I have not exposed a way to mutate the effect classes after creation. The cost of creating an effects class is not high, but they should be disposed to avoid having them clutter the finalizer queue.

I also did not include the red eye removal effect as it is sort of a one-off from general image effects. GDI+ won't be adding any new effects (these were added over 20 years ago).

Tests are pending.