Add convenience / performance related methods to ListViewItemCollection / ControlCollection
ListViewItemCollection doesn't have an AddRange that takes an IEnumerable<ListViewItem>, which necessitates unnecessary array allocations when adding from a collection.

It also doesn't implement a generic IList<T> which makes typed usage and LINQ usage difficult. This adds that.

Also add IEnumerable<Control> to ControlCollection to address the LINQ scenario. Control has IList, but indexed setting throws. I don't want to add IList<Control> until we've had time to evaluate the implications of implicitly doing the public steps with the collection that replicate what those setters should be doing.

Also tweak ArrangedElementCollection so the static empty collection can't be written to.

This doesn't fix the ambiguity of `[..]`, but it does avoid the need for it in some cases.