Clean up initialization messages
This change cleans up some of the instances of the following analyzer messages (using the code fixes primarily):

- Simplify type names (IDE0002)
- Redundant casts (IDE0004)
- Collection initialization syntax (IDE0028)
- Simplify object initialization (IDE0017)
- More specific types for performance (CA1859)

There are issues with applying initialization syntax to test theory data so I haven't touched much of the test code. Using collection initialization also doesn't work with some of our collections (such as DesignerVerbCollection). I'm pretty sure that is related to them not implementing a `IEnumerable<T>`.

The plan is to follow up and make sure our collections have at least `IEnumerable<T>` to enable using this syntax and that using said syntax doesn't go down the `Add(object)` path if said overload exists.

Additionally I hope to turn some of these into at least warnings.

The hope is that with less for the analyzers to report it will improve the development experience in VS (and it appears to).