Finish cleaning member naming
This finishes addressing member names to allow making the analyzer a failure.

There is still room for improvement on some of these member names (notably abbreviations).

This change also makes specifying accessiblity modifiers an error and warns about adding the readonly modifier (not an error to facilitate iterating in VS).