Remove unused code (#11057)
This enables a number of additional analyzers. The key one is turning unused code into warnings. Ends up we had quite a bit of unused parameters and fields.

I've tried to err on the side of caution when properties or methods might be used in the designer.

This also enforces using switch expressions when possible.

Other enablements (that didn't change much):

- Use keywords for types (int over System.Int) is now a warning
- Use null propagation is  now a warning
- Precompiled regex is now an error
- No protected members in sealed types
- Suppression format checks