Remove finalization watch from System.Drawing (#11061)
Removing code that does not compile as part of our normal builds.

In the future we can add debug code in the finalizer methods that will fail when the finalizer is hit (as we do with the scopes for HDC, etc.). Keeping an initialization stack isn't super useful.

Note that the try catch blocks are not needed as with direct interop we shouldn't see exceptions that we would handle.