Add a general deserializer for BinaryFormattedObject (#11272)
This adds a constrained, simplified deserializer base on BinaryFormattedObject.

This deserializer does not support:

- IObjectReference
- Offset arrays

The deserialzer guarantees that value type members are initialized in SerializationInfo callbacks and deserialization events. It will throw if value types can not be applied in their final state.

Reference type members (including transitive ones) are only guaranteed to be in their final state if the object graph has no cycles (as is also the case for BinaryFormatter).