BinaryFormat test updates
The most expensive thing we're doing in the binary format tests is getting the validation test method for types. I've added a cache for this, which cuts a third of the testing time for the project.

Also adding some theory data helper classes for tuples for clarity and minor perf.

Finally, no need to call the slow method for array index sets for multidimensional arrays.