Some GDI+ Font cleanup
GDI+ installed FontCollection is a static. We reference that indirectly in a few places, but are not consistent about it. This comes into play with both the FontCollection and FontFamily, as FontFamily instances from the installed FontCollectionare also "effectively" static.

This does some cleanup, with additional changes to follow. Notably:

- Minor tweaking of GdiCache for naming and to move a TextRenderer specific helper to TextRenderer
- Move a define down to Core
- Clean up FontCollection to encapsulate setting the pointer
- Suppress finalization on InstalledFontCollection as it does nothing
- Make an internal instace of InstalledFontCollection
- Don't allocate a string for FontFamily.GetHashCode
- Don't allocate a string to strip off the @ on the Font name when constructing a FontFamily
- Remove debug instance count on FontFamily

The key follow up change here will be to properly ref count FontFamily. It isn't being done right, which has an impact on PrivateFontCollection's ref counting. (Ref counting a FontFamily only impacts the FontCollection it came from. FontFamily objects aren't deleted directly. They are indirectly deleted when the FontCollection has reached 0 ref count.)