Ensure source paths leave 40 characters for root
The default path for VS is in the user directory, which ends up being something like `C:\Users\jkuhne\source\repos\winforms`.

Recent paths make the repo impossible to sync. I've added a test that reserves 40 characters for the repro root path and moved files out of subdirectories that violated that.

A longer follow up is to see if we can pick some type names that aren't so long. Additionally we should try to up this check to 50 characters- which would require more renaming.