Stop locking internal objects with `this` (#11842)
Better to have explicit lock objects from a maintenance perspective.

Ideally we can whittle down some of the other locks on non Lock objects, but we need to be very careful about where we might be inadvertently exposing the ability to lock externally. Even if you can technically do so, we might want to take the breaking change for some of these in 10.