Fix:The SplitContainer control cannot be deleted successfully under ToolStripContainer tab #11793 (#11830)
## Proposed changes
When SplitContainer is removed, ComponentDesigner, from which SplitContainerDesigner is derived, will set it's property `Component` null.

And there is a `ComponentRemoveHandler` which access SplitContainerDesigner.ActionLlists. 

SplitContainerDesigner.ActionLlists will create OrientationActionList, in turn, OrientationActionList will access ComponentDesigner.Component.

But ComponentDesigner.Component will throw Exception when Component is null.

So modify the logic when create SplitContainerDesigner.ActionLlists, we first check if ComponentDesigner.Component is null.