belated feedback on https://github.com/dotnet/winforms/pull/11835/ (#11882)
This test compares length of file paths in this repo against the MAX_PATH const, we assume that path length of the repo root is 40 chars or less, this leaves MAX_PATH - 1 - 40 chars for the path that we see in the GH (portion from src to .cs). This test enforces this limit be enumerating all repo files, getting the full path length and comparing this length against the maximim. Maximum full path length should be compensated for repo roots that are shorter than 40.

MaxRootLength + <relative path to the file>.Length < 259
MaxRootLength + <full path to the file>.Length  - currentRootLength  < 259
<full path to the file>.Length  < 259 - MaxRootLength + currentRootLength