Move child controls collection out of PropertyStore (#12191)
It is something we look at constantly when we have it, so I've moved it into an internal auto property. Ths avoids some overhead and complexity.

I've also changed more of the depreciated PropertyStore methods to the new overloads.

Removed the repeated perf note. One should always assume that explicitly using indexing to iterate is intentional.

We really shouldn't allow controls to return null accessible objects. Check for this and remove test that was doing it.