Add internals visibility to System.Windows.Primitives
Need a bottom level interop assembly in WPF to give all projects a common base so that CsWin32 doesn't generate duplicate defines.

WindowsBase might be able to subsume this at some point, but it doesn't have internals visible to the rest of the graph and it's current internal interop is included repeatedly in multiple assemblies.