Tweak visibility for WPF (#12276)
Added the full set of runtime assemblies with (hopefully) the right keys. Test assemblies will eventually need added.

I've introduced constants for the public key values to make this a little easier to update / follow. I didn't update all of the usages, anyone else can do so, or I'll eventually pick it off when it is convenient.

This also moves some interop down to Core and does some renaming. It is intentional that I did not change the name of some of the files to allow history to be easily kept. I'll update names to match in future pulls.