Tweak analyzers
Want to keep false positives to a minimum, even in "Messages".  The var message was coming up a lot when editing files where we wouldn't want to "fix" it, so I've silenced it.

There are a lot of "if statement can be simplified" messages in our files. I went through and fixed a bunch that came up when I was working on other things to validate the usefulness. It is the better choice at least 80% of the time, so I'm leaving it on.

Changes were primariliy made using the automated fixes. Flipped a few conditional blocks, again with the tooling.