Remove some dead conditional code (#12451)
CA1508 finds these, but it has many false positives from native calls and Debug asserts.

I've fixed a batch of real issues and tweaked some code to not get false positives. Won't be able to turn this one on by default, but I'll make a few more passes to find real places where there are issues. (There are over 200 hits.)

Considering making this one a suggestion, but still leaning towards no because of the false positives.