Re-enable some analyzers
There were a few marked as TODO. This makes them back into warnings with the exception of IDE0059 which is "Unnecessary assignment of a value". That is now a suggestion. Some of the fixes for IDE0059 are in this commit, others will follow as there are a lot of failures on that one.