Move all GDI+ specific interop to System.Private.Windows.GdiPlus (#12633)
This removes the risk of not properly initializing GDI+ before calling native methods. It also removes the risk of inadvertently loading GDI+ in WPF. Lastly, it reduces the footprint for WPF.

The new project is not under a second `src` folder as we're going to move all tests to the root under `tests`. This allows for better build file structure (Directory.*, etc.) and helps some with MAX_PATH constraints.