[release/8.0] Fix #12661 Null Reference Exception: System.Windows.Forms.TabControl<WmSelChange> (#12731)
Fixes 12661
Customer Impact
A customer reported a crash in their application that uses WinForms TabControl and requested a fix in NET8 release. Application is running under accessibility tools and is removing TabPages from a control dynamically. When the selected page is removed, the selection moves to the next page and the corresponding accessibility event is queued using BeginInvoke to be raised. If by the time the event is raised, the page is removed, a NullReferenceException is thrown. No workaround is available.

Fix
Test if the selected page is still available before raising the accessibility event inside the BeginInvoke delegate.