Refactor DataFormats
Prepare DataFormats to move core implementation to share with WPF.

This splits the format list into two collections. The first is the known predefined format ids that come from Windows headers. The second collection contains ones that must be looked up. This optimizes allocations for typical user scenarios in WinForms and WPF.

An internal interface is used to abstract the Format class.

Next PR will move the shared logic into the shared assembly.