Add hollow types for deprecated UI controls (#10913)
Fixes #3783

Add controls removed in NET3.1 Back to .NET10 for binary compatibility.

None of the re-added types can be constructed, all constructors throw new PlatformNotSupportedException(), default constructors are suppressed by introducing throwing public constructors if needed.
All attributes that have been present on the .NET Framework types, are preserved, as they are accessible by reflection.
All re-introduced types are decorated with [EditorBrowsable(EditorBrowsableState.Never)] attribute that hides them from the intellisense. Type names can be typed in by the developer manually and intellisense would show the same members as it did for .NET Framework projects.
All re-introduced types are decorated with the ObsoleteAttribute that results in a compile time warning. All types share a single deprecation warning Id, to simplify suppression, but have different explanation messages.
All re-introduced types are decorated with [Browsable(false)] attribute to not show custom control properties of these types in the property browser.
All public members are re-introduced with their metadata, except for the private attributes.
Members inherited from the base classes (Control, for example) are not re-introduced even if they had been overridden in the past because they are not required for binary compatibility, unless they are decorated with different attributes.
Members that are re-added to the existing types (Form and Control) are returning defaults or doing nothing.
Non-void public or protected instance members on the restored types throw null, not the PlatformNotSupportedException to save space.
Void public or protected instance members do nothing.
Public or protected fields return default values.
Nullability is disabled for all re-added classes, structs and delegates for compatibility with the .NET Framework.